Update README.md
For this project, I created tables to organize data for Pewlett Hackard's (a fictional company) research on employees from the 1980s and 1990s. We relied solely on six CSV files, which were all that remained of the employee database from that period. The project involved performing data modeling, data engineering, and data analysis.